requirements.txt
beancount
camelot-py
pandas
tabula-py
python-docx
pdftotext
python-magic